backend/pyproject.toml
[project]
name = "thesisv2"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aiortc>=1.13.0",
    "alembic>=1.16.5",
    "anyio>=4.9.0",
    "authlib>=1.6.4",
    "azure-cognitiveservices-speech>=1.47.0",
    "deepeval>=3.6.9",
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.116.1",
    "httpx>=0.27.0",
    "langchain>=1.0.2",
    "langchain-azure-ai>=1.0.0",
    "langgraph>=1.0.1",
    "numpy>=2.3.1",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "python-ulid>=3.0.0",
    "scipy>=1.16.0",
    "scipy-stubs>=1.16.1.0",
    "sqlalchemy>=2.0.43",
    "streamz>=0.6.4",
    "tqdm>=4.67.1",
    "vosk>=0.3.45",
    "websockets>=15.0.1",
]

[tool.basedpyright]
include = ["src/"]
exclude = [
    "**/__pycache__",
]

[[tool.basedpyright.executionEnvironments]]
root = "backend/src"

# Ensure we don't run llm tests (expensive)
[tool.pytest.ini_options]
addopts = "-m \"not llm\""
markers = [
    "llm: marks tests that use LLM evaluation (run with `uv run deepeval run test`)",
]

[dependency-groups]
dev = [
    "basedpyright>=1.32.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.7",
]

[tool]


[tool.alembic]

# path to migration scripts.
# this is typically a path given in POSIX (e.g. forward slashes)
# format, relative to the token %(here)s which refers to the location of this
# ini file
script_location = "%(here)s/alembic"

# template used to generate migration file names; The default value is %%(rev)s_%%(slug)s
# Uncomment the line below if you want the files to be prepended with date and time
# see https://alembic.sqlalchemy.org/en/latest/tutorial.html#editing-the-ini-file
# for all available tokens
# file_template = "%%(year)d_%%(month).2d_%%(day).2d_%%(hour).2d%%(minute).2d-%%(rev)s_%%(slug)s"

# additional paths to be prepended to sys.path. defaults to the current working directory.
prepend_sys_path = [
    "./src"
]

# timezone to use when rendering the date within the migration file
# as well as the filename.
# If specified, requires the python>=3.9 or backports.zoneinfo library and tzdata library.
# Any required deps can installed by adding `alembic[tz]` to the pip requirements
# string value is passed to ZoneInfo()
# leave blank for localtime
# timezone =

# max length of characters to apply to the "slug" field
# truncate_slug_length = 40

# set to 'true' to run the environment during
# the 'revision' command, regardless of autogenerate
# revision_environment = false

# set to 'true' to allow .pyc and .pyo files without
# a source .py file to be detected as revisions in the
# versions/ directory
# sourceless = false

# version location specification; This defaults
# to <script_location>/versions.  When using multiple version
# directories, initial revisions must be specified with --version-path.
# version_locations = [
#    "%(here)s/alembic/versions",
#    "%(here)s/foo/bar"
# ]


# set to 'true' to search source files recursively
# in each "version_locations" directory
# new in Alembic version 1.10
# recursive_version_locations = false

# the output encoding used when revision files
# are written from script.py.mako
# output_encoding = "utf-8"

# This section defines scripts or Python functions that are run
# on newly generated revision scripts.  See the documentation for further
# detail and examples
# [[tool.alembic.post_write_hooks]]
# format using "black" - use the console_scripts runner,
# against the "black" entrypoint
# name = "black"
# type = "console_scripts"
# entrypoint = "black"
# options = "-l 79 REVISION_SCRIPT_FILENAME"
#
# [[tool.alembic.post_write_hooks]]
# lint with attempts to fix using "ruff" - use the module runner, against the "ruff" module
# name = "ruff"
# type = "module"
# module = "ruff"
# options = "check --fix REVISION_SCRIPT_FILENAME"
#
# [[tool.alembic.post_write_hooks]]
# Alternatively, use the exec runner to execute a binary found on your PATH
# name = "ruff"
# type = "exec"
# executable = "ruff"
# options = "check --fix REVISION_SCRIPT_FILENAME"

